Central Widget
To prevent opening new windows all the time, we use the Central Widget
of the Mainwindow to show the different functionalities of the program.

Still buggy: if a functionality is 'closed', we still have to show the
original buttons so that the user can continu using the program. (now it
just shows a blank screen, not useful)