class KiesOphaalpunten
in class NieuweAanmelding, the user can input calls of locations that
the truck can visit on a new route. After collecting these calls, the
user can use the class KiesOphaalpunten to select those locations that
might be feasible for the next trip of the truck (keeping in mind the
maximum load of the truck)

This starts to get interesting :-)

For now, the classdoesn't make an actual connection to the database yet,
but uses dummy input (pieces of fruit, based on the example of
QListWidget). Hope to work this out further tomorrow. Good night!