Transportationlist: no filename asked, just date
1. it is impossible to leave a QDateEdit empty, so we take the date of
yesterday as a standard. If the QDateEdit < currentDate, then we know
the date was not changed (or the user wants to generate a
transportationlist in the past). The user will be asked whether to
proceed or not.

2. the filename of the transportationlist will be ["vervoerslijst"] +
[DATE] + [UUID] + [".doc"] , the filename of the map will be ["kaart"] +
[DATE] + [UUID] + [".png"]

Could be made more "human readable" in a future version, but will do for
now.