Correcting SpinBox->Maximum, the legend continues: 3
By default, a QSpinBox can only handle up to 99. This is too small for
the transportationlist and the error was viewable as any higher number
was mentioned as '99' in the documents for the client. Whoopsie...

This time, also corrected for the 'bags' when new storage amounts are
being announced (nieuweaanmelding.cpp)