Adding column 'aanmeldingsdatum' to collection history
In the database, the table "ophalinghistoriek" now has an extra column
"aanmeldingsdatum" which will contain the date that the location
reported the volume/weight of their stock.

Up till now, this information was kept in table "aanmelding", but got
deleted once a collection route was being confirmed.

See mail with request, dated November 3rd, 2014