Ignoring locations from the confirmation module is now possible (locationShouldBeSkippedCheckBox)
Regarding_locationShouldBeSkippedCheckBox:

It annoys me that I can't use ->checkState, but it works when I check
the ->text() for the string "true"
(see below)

What is going wrong?

For now it works, but it would be cleaner if we can simply check the
state of the locationShouldBeSkippedCheckBox ?

In initiating the mapper, maybe the following can be used:
 // mapper->addMapping(locationShouldBeSkippedCheckBox, 8, "checked");

Or is it because of Flag-settings? Look at [
listofophaalpuntentocontact.cpp ] or [ kiesophaalpunten.cpp ] ?
 // item->setFlags(item->flags|Qt::ItemIsUserCheckable);
and what is the difference with [ item->SetCheckable() ] ???