Google Maps JavaScript API v3 release instead of experimental
'geocodezip' remarks in
http://stackoverflow.com/questions/27730642/google-maps-draw-route-
of the Google Maps JavaScript API v3.

I am actually not convinced, I believe I have to request the
experimental version explicitly through adding "v=3.exp", but I do not
want to risk it.

What I do have noticed, is that the "sensor" variable has become
obsolete, so I deleted that one, and in HTML5 you do not need to state a
script-type as 'text/javascript' anymore, so that has been removed as
well.

Still, the bug as I explain in the StackOverflow mentioned above, is
still not solved with these adaptations.
Breaking my teeth on it since 31st of December and haven't dared to
commit since then. But this small adaptation should be correct.

Help is welcome though...