First attempt Adding Javascript: [FAILED]
Error:
   No bp log location saved, using default.
   [000:000] Cpu: 6.15.11, x2, 1795Mhz, 1985MB
   [000:002] Computer model: Not available
   [000:002] Browser XEmbed support present: 1
   [000:002] Browser toolkit is not Gtk2 (0).
   [000:004] Using Xt toolkit
   No bp log location saved, using default.
   [000:000] Cpu: 6.15.11, x2, 1795Mhz, 1985MB
   [000:000] Computer model: Not available

I suppose this is an error log from the WebKit. I have no solution yet,
but I suppose it is because I am adding a listener to _every_ marker,
which becomes too much.

Maybe better to add a listener to the map, then everytime we get a
double click on the map, we can check whether it is a marker or not.

Reading:
 * http://stackoverflow.com/questions/6611634/google-maps-api-v3-add-
 * http://stackoverflow.com/questions/9944585/adding-thousands-of-
 * https://developers.google.com/maps/articles/toomanymarkers?hl=en