QListWidget cannot signal a Drag/Drop
this is annoying, as when the ophaalpunten have changed order in the route, of course the total distance / total time spent will be different, but the routetool has no clue that he has to recalculate.
Now we found a way to spot that a Drag/Drop was done since the last time we calculated the total distance / time spent.
(when the order of ophaalpunten in the QListWidget is different from the order in m_markers (copy in our memory that we use to calculate with), then a Drag/Drop occured and we first have to bring m_markers back up to date (using Form::reorderMarkers() )