Bug-fix: when we input a delivery, it doesn't get found by Google Maps. Next, if we input an ordinary address, it finds the address and marks it on the map, but with the name (and address) of the delivery
problem was: when inputting a delivery, GeocodeDataManager::pushLevering(SLevering levering) gets called and sets the variable marker_type to Levering.
When that lookup gives ZERO_RESULTS, nobody resets the variable.

It gives no problem when you input a Ophaalpunt, because that one has its own function pushListOfMarkers(QList<SOphaalpunt> *) which will set the marker_type to Ophaalpunt

Instead of resetting the variable when an error occurs, we just make sure that the variable is correctly set to Address when an ordinary address has been input.