Editing Routes (9) prepare UI: notice user he is editing route
a box appears at the top of the routetool, informing the user he is
currently editing a planned route.

To catch the eye, we made the text blue (and a cancel button next to it)

issue #40: the database only saves 'aanmeldingen', but no deliveries nor 'simple addresses'. This should be fixed in the future