Show all locations: Javascript gets in trouble with //' !
On 31st of March 2016, the user informed us that the button "show all locations" no longer functioned.
Immediately, we thought of:
 * Google Maps change? (but impossible, it worked in the Linux-test environment)
 * browser update? (as the user runs on Windows environment)
 * Qt-version difference?

Appears to be a problem with escaping the names of locations before running a JavaScript that shows all the markers. After long reasearch, we found at that location number 18 is called "Commune d\'Etterbeek" and that extra slash before the ' is of major significance.
Our script would replace every ' by a \' . In case of this location, this would result in "Commune d\\'Etterbeek", making the slash escaped but not the ', and therefore we got a JavaScript error which resulted in not showing any marker.

Now we have added this functionality so that, after escaping the location names and addresses, if a \\' is found, it will be replaced by \' .